refactor: create folder for legend and rename, consolidate, and move files for legend and hovering related code
