+++ giraffe/src/components/BandHoverLayer.tsx
@@ -5,12 +5,12 @@ import {Tooltip} from './Tooltip'
 import {Props as BandLayerProps} from './BandLayer'
 import {FILL} from '../constants/columnKeys'
 import {LineHoverDimension, LineData} from '../types'
-import {getBandTooltipData} from '../utils/tooltip'
-import {getBandHoverPoints} from '../utils/getBandHoverPoints'
+import {getBandTooltipData} from '../utils/legend/tooltip'
+import {getBandHoverPoints} from '../utils/bandHover'
 import {drawLines} from '../utils/drawLines'
 import {drawLineHoverData} from '../utils/drawLineHoverData'
 import {useCanvas} from '../utils/useCanvas'
-import {BandHoverIndices} from '../utils/getBandHoverIndices'
+import {BandHoverIndices} from '../utils/bandHover'
 import {isNumber} from '../utils/isNumber'
 
 interface Props extends BandLayerProps {
