+++ giraffe/src/components/LineHoverLayer.tsx
@@ -5,8 +5,8 @@ import {Tooltip} from './Tooltip'
 import {Props as LineLayerProps} from './LineLayer'
 import {FILL} from '../constants/columnKeys'
 import {LineHoverDimension, LineData} from '../types'
-import {getPointsTooltipData} from '../utils/tooltip'
-import {getLineHoverPoints} from '../utils/getLineHoverPoints'
+import {getPointsTooltipData} from '../utils/legend/tooltip'
+import {getLineHoverPoints} from '../utils/lineHover'
 import {drawLines} from '../utils/drawLines'
 import {drawLineHoverData} from '../utils/drawLineHoverData'
 import {useCanvas} from '../utils/useCanvas'
