+++ giraffe/src/components/ScatterHoverLayer.tsx
@@ -4,7 +4,7 @@ import {ScatterLayerConfig, ScatterLayerSpec, LayerProps} from '../types'
 import {useHoverPointIndices} from '../utils/useHoverPointIndices'
 import {drawPoints} from '../utils/drawPoints'
 import {Tooltip} from './Tooltip'
-import {getPointsTooltipData} from '../utils/tooltip'
+import {getPointsTooltipData} from '../utils/legend/tooltip'
 import {SCATTER_HOVER_POINT_SIZE} from '../constants'
 import {FILL, SYMBOL} from '../constants/columnKeys'
 import {useCanvas} from '../utils/useCanvas'
