+++ giraffe/src/components/Tooltip.tsx
@@ -3,7 +3,7 @@ import {FunctionComponent, useMemo} from 'react'
 import {createPortal} from 'react-dom'
 
 import {TooltipData, Config} from '../types'
-import {useTooltipElement} from '../utils/useTooltipElement'
+import {useTooltipElement} from '../utils/legend/useTooltipElement'
 import {TOOLTIP_MAXIMUM_OPACITY, TOOLTIP_MINIMUM_OPACITY} from '../constants'
 
 import {generateTooltipStyles, TooltipDotsStyles} from './TooltipStyles'
