+++ giraffe/src/utils/bandHover.ts
@@ -1,5 +1,37 @@
-import {NumericColumnData, BandIndexMap, LineData} from '../types'
+import {NumericColumnData, BandIndexMap, LineData, Scale, Table} from '../types'
+// import {LineData, } from '../types'
 import {isDefined} from './isDefined'
+
+import {FILL} from '../constants/columnKeys'
+import {isNumber} from './isNumber'
+
+export const getBandHoverPoints = (
+  table: Table,
+  hoverRowIndices: number[],
+  xColKey: string,
+  yColKey: string,
+  xScale: Scale<number, number>,
+  yScale: Scale<number, number>,
+  lineData: LineData
+): Array<{x: number; y: number; fill: string}> => {
+  const xColData = table.getColumn(xColKey, 'number')
+  const yColData = table.getColumn(yColKey, 'number')
+  const groupColData = table.getColumn(FILL, 'number')
+  const lines = lineData || {}
+
+  return hoverRowIndices.map(hoverIndex => {
+    let color = ''
+    const lineIndex = groupColData[hoverIndex]
+    if (isNumber(lineIndex) && lines[lineIndex]) {
+      color = lines[lineIndex].fill
+    }
+    return {
+      x: xScale(xColData[hoverIndex]),
+      y: yScale(yColData[hoverIndex]),
+      fill: color,
+    }
+  })
+}
 
 interface LineLengths {
   [index: string]: {
