+++ giraffe/src/utils/lineHover.ts
