+++ giraffe/src/utils/useBandHover.ts
